feat: flyway 세팅, id 값 long -> int (#20)
* feat: flyway 세팅, id 값 long -> int

* rebase & 테스트에서 DDL 제거, DML 리셋

---------

+++ api/src/main/kotlin/siksha/wafflestudio/api/controller/BoardController.kt
@@ -17,7 +17,7 @@ class BoardController(
 
     @GetMapping("/{board_id}")
     fun getBoard(
-        @PathVariable("board_id") boardId: Long,
+        @PathVariable("board_id") boardId: Int,
     ): Optional<BoardDto> = Optional.ofNullable(boardApplicationService.getBoardById(boardId))
 
     @PostMapping
