+++ api/src/main/kotlin/siksha/wafflestudio/api/controller/CommentController.kt
@@ -25,7 +25,7 @@ class CommentController(
     @GetMapping("/community/comments/web")
     fun getCommentsWithoutAuth(
         request: HttpServletRequest,
-        @RequestParam(name = "post_id") postId: Long,
+        @RequestParam(name = "post_id") postId: Int,
         @RequestParam(name = "page", defaultValue = "1") @Min(1) page: Int,
         @RequestParam(name = "per_page", defaultValue = "10") @Min(1) perPage: Int,
     ): GetCommentsResponseDto? {
@@ -35,11 +35,11 @@ class CommentController(
     @GetMapping("/community/comments")
     fun getComments(
         request: HttpServletRequest,
-        @RequestParam(name = "post_id") postId: Long,
+        @RequestParam(name = "post_id") postId: Int,
         @RequestParam(name = "page", defaultValue = "1") @Min(1) page: Int,
         @RequestParam(name = "per_page", defaultValue = "10") @Min(1) perPage: Int,
     ): GetCommentsResponseDto? {
-        return commentService.getComments(request.userId,postId, page, perPage)
+        return commentService.getComments(request.userId, postId,  page, perPage)
     }
 
     @PostMapping("/community/comments")
@@ -54,7 +54,7 @@ class CommentController(
     @PatchMapping("/community/comments/{commentId}")
     fun patchComment(
         request: HttpServletRequest,
-        @PathVariable commentId: Long,
+        @PathVariable commentId: Int,
         @RequestBody patchDto: PatchCommentRequestDto,
     ): CommentResponseDto {
         return commentService.patchComment(request.userId, commentId, patchDto)
@@ -64,7 +64,7 @@ class CommentController(
     @ResponseStatus(HttpStatus.NO_CONTENT)
     fun deleteComment(
         request: HttpServletRequest,
-        @PathVariable commentId: Long,
+        @PathVariable commentId: Int,
     ) {
         commentService.deleteComment(request.userId, commentId)
     }
@@ -73,7 +73,7 @@ class CommentController(
     @ResponseStatus(HttpStatus.CREATED)
     fun createCommentLike(
         request: HttpServletRequest,
-        @PathVariable commentId: Long,
+        @PathVariable commentId: Int,
     ): CommentResponseDto {
         return commentService.createOrUpdateCommentLike(request.userId, commentId, true)
     }
@@ -83,7 +83,7 @@ class CommentController(
     @ResponseStatus(HttpStatus.CREATED)
     fun createCommentUnlike(
         request: HttpServletRequest,
-        @PathVariable commentId: Long,
+        @PathVariable commentId: Int,
     ): CommentResponseDto {
         return commentService.createOrUpdateCommentLike(request.userId, commentId, false)
     }
@@ -92,7 +92,7 @@ class CommentController(
     @ResponseStatus(HttpStatus.CREATED)
     fun createCommentReport(
         request: HttpServletRequest,
-        @PathVariable commentId: Long,
+        @PathVariable commentId: Int,
         @RequestBody createDto: CreateCommentReportRequestDto,
     ): CommentsReportResponseDto {
         return commentService.createCommentReport(request.userId, commentId, createDto.reason)
